ci: Make all changes eligible for versioning and release (#70)
Markdown changes had previously been ignored by lerna, but this was causing issues with the changelog when docs were updated in packages.

Updates to tests and stories would encounter the same issues.

This change makes it so that the release step will now depend on the conventional commit alone to determine appropriate bumps, and group all previous unreleased changes (e.g. updates to tests, refactors) in the next package release.