BREAKING CHANGE: Add generic elm modal
Elm requires directories to be capitalised when resolving elm modules.

In order to satisfy this the 'Modal/primitives' directory has been
renamed to 'Modal/Primitives'.

This breaks all current calls to 'Modal/primitives'.

Internally this has been updated for kaizen but will need to be updated
for other services using 'Modal/primitives'.