feat: Add confirmation variant to elm modal
Move modalBox rendering to GenericModal module

The modal box belongs to the generic modal so it makes sense that the
styles and view be defined in that module.

Add onDismiss to ModalHeader config

Providing an onDismiss will ensure the modal header renders a close
button, otherwise it will omit rendering the close button and consumer
will need to handle closing.

Add confirmation variant

Add Confirmation modal view

Add inline styles for width and height

The elm modal differes slightly form the react version in how it determines its sizes.
The generic modal is the base view for all modal variants and the width and
heights are constrained here.

Add size type for GenericModal

This creates sensible defaults for modal width as well as a Custom
variant that can receive whatever width and height the consumer sets.

From a responsiveness point of view it may not be a great idea to allow
custom widths and heights. This can defenitely be revised.

Use scroll layer instead of modalLayer

Because elm modal views are constructed slightly differently, it does
not need to use a modalLayer.