feat: Render a close icon when onDismiss is present in ModalHeader
ModalHeader can be passed a view where the close button is accounted
for, or the consumer can just pass the onDismiss msg and ModalHeader
will render a close button which dispatches that onDismiss message.

When ModalHeader receives an onDismiss property it will position a close
button on the view.

The consumer may like to provide their own close button when passing the
header content to the ModalHeader.