feat: Add confirmation variant to elm modal (#168)
* feat: Add confirmation variant to elm modal

Move modalBox rendering to GenericModal module

The modal box belongs to the generic modal so it makes sense that the
styles and view be defined in that module.

Add onDismiss to ModalHeader config

Providing an onDismiss will ensure the modal header renders a close
button, otherwise it will omit rendering the close button and consumer
will need to handle closing.

Add confirmation variant

Add Confirmation modal view

Add inline styles for width and height

The elm modal differes slightly form the react version in how it determines its sizes.
The generic modal is the base view for all modal variants and the width and
heights are constrained here.

Add size type for GenericModal

This creates sensible defaults for modal width as well as a Custom
variant that can receive whatever width and height the consumer sets.

From a responsiveness point of view it may not be a great idea to allow
custom widths and heights. This can defenitely be revised.

Use scroll layer instead of modalLayer

Because elm modal views are constructed slightly differently, it does
not need to use a modalLayer.

* feat: Render a close icon when onDismiss is present in ModalHeader

ModalHeader can be passed a view where the close button is accounted
for, or the consumer can just pass the onDismiss msg and ModalHeader
will render a close button which dispatches that onDismiss message.

When ModalHeader receives an onDismiss property it will position a close
button on the view.

The consumer may like to provide their own close button when passing the
header content to the ModalHeader.

* Rename presets folder to be elm friendly

BREAKING CHANGE: Elm requires modules to be resolved with capitalized folder names.

* chore: Build elm story for Confirmation Informative modal

* Add title to confirmation config

* Render body on confirmation modal

* feat: Make background color flexible on ModalBody

Some modal bodies may want to have a stone background instead of the
default white. This prop allows the consumer both options.

Previously ModalBody background color was set to stone, now by default
it is white.

* Add ModalFooter to confirmation modal

* Add onDismiss msg to cancel button

* Add onConfirm and onDismiss msgs to confrimation modal

Remove styling

This styling makes more sense being done in the ModalFooter component

Add style to each action button in list

This styling differes slightly from the react version as it does not
row-reverse the action buttons.

* rename ModalBody scss file

There is no need to namespace this scss file with .elm

Clean up some css class definitions

* Add confirm and dismiss label configuration

* Add padding to Confirmation config

Change argument name

The size being passed in is not a maybe its a Size type