feat: Add `boxOffset` property to the `Popover` component (#181)
Add `boxOffset` property to the `Popover` component. It will shift the popover box a specified amount, without moving the popover arrow.

![image](https://user-images.githubusercontent.com/4495057/71139166-81054600-2261-11ea-8e8d-8e6133d03e0f.png)

*(the snapshot above is slightly off centre, but that has now been fixed)*

This task links to a requirement of ticket CORE-348 in perform. We needed to show a popover, have the popover arrow centered to the sibling button, but the box to not go past the main body margin. This new property will give us the flexibility to achieve this.

Side updates:
* Removal of some dead scss code in the `arrow` mixin
* Removal of the `position` `left` and `right` options, since they aren't yet implemented.
* Separate the popover arrow to its own element, as to give more flexibility on how it is positioned.

Notes:
One of the ugly things that I've introduced is some fragmentation of styling. Because we now allow for manually updating the position via javascript, we are forced to have some styling in javascript, and some in scss. I'm not sure the best way to get around this, other than going all in on javascript styles (which I would be very happy to do 😛).

I've gone through all of the storybook stories, to make sure there are no visual regressions, and it all looks good on my end.