use the Storybook 5.3 main.js format for finding stories
it seems necessary for CSF to work.
note that this path is much less flexible than the current one:
it only looks for stories in packages/component-library/
I couldn't get the regex or glob to work in the new format, so this will
fixed in a later release. Currently this is the only directory where stories are found.