feat: Elm modal focus first focusable element (#277)
* feat: Elm modal focus first focusable element

Focus the default focusable element when the modal opens.

This takes the focus away from the trigger and into the modal itself.

-------------

This PR sets up the ground work to implement a full focus management
system that includes focus lock, and returning focus to trigger.

-------------

Seperate modal data from modal state

Modal data is not in anyway dependant on modal state. Modal data is just
information about the modal that is passed around.

Because there is no dependancy it makes more sense to decouple both
types and instead return a tuple of (modalState, modalData).

-------------

Set ids on focusable elements in Confirmation modal

To manage focus in confirmation variants we need focusable elements to
have ids that can be targeted when setting focus.

-------------

Setting focus on default focusable element

This includes the case where the modal is forced open via the helper. In
this case subscriptions trigger the focus of the default element.

In the case where the default element is not found Modal tries to
fallback to the last focusable element.

* set force open to trigger animationFrame

* Remove unecessary css class

The scrollLayer resolves to overflow-y: auto; so no need to set with new
class.