feat: Allow for inputs to be uncontrolled (#279)
Co-authored-by: Seb Pearce <sebpearce@gmail.com>

This PR allows for the `Input` and `TextField` components to become uncontrolled. The `inputValue` and `onChange` properties have now become optional, and the `inputRef` property has been added, as so we can imperatively query the textfield value.

Don't get me wrong, I love the idea of deriving things like validation from state, but the downside with controlled components is that they come with a performance cost.

I am currently investigating the [`react-hook-form`](https://react-hook-form.com/) library as potential option for perform, which will not work unless we allow for uncontrolled components in kaizen. Regardless if we end up taking this path, I think it would still be a good thing to provide the choice of controlled and uncontrolled components in kaizen. If this PR gets rejected, I'l scrap the idea of using react-hook-form.