refactor: Make elm modal completely dependant on subscriptions (#283)
* refactor: Make elm modal completely dependant on subscriptions

BREAKING CHANGE: The modal has been changed to exclusively work with
subscriptions which massively siimplifies the modal logic.

As a result using the trigger function simply updates the ModalState
which fires the subscriptions.

This allows timing to be done with a requestAnimationFrame instead of a
Process.sleep. This change removed a lot of the jank the modal was
experiencing.

Changing implementations is very easy, remove Cmds and status returns
from the trigger function. It simply just returns ModalState.

----------

Instead of returning Cmd ModalMsgs when triggering the modal it makes
more sense to simply ask the modal to run. This is a simple field in the
modal data that will be picked up by subscriptions.

This will then trigger the modal update in which case the modal can do
all the normal things it was doing previously. This moves all the modal
running logic to just one flow.

----------

* lazy load main view

This provides a pretty significant performance boost. The animation
frame absolutely hammers the view without it. Because no data changes
during the animation frame this avoids a redraw of the view.