docs: Fix sass module resolution issue in package docs (#304)
There's a bug where css-loader will use node's package resolution strategy for scss imports, and so follows the `main` key in the package.json for imports which only specify the package name. This results in the published javascript being parsed as css during builds in consuming projects.

This change updates the API docs for this package to specifically recommend importing the scss file.