feat: Add InputEditModal variant for Elm Modal
Create a reusable accessible title function

This acts as a utility function that ensures an ariaLabelledBy id lives
on the parent div to the heading child

Add helper function for ariaDescribedBy.

This function adds the ariaDescribedBy to the id of the containing div
which surrounds the instructive text or subtext of the modal body.