Set up modal stories to set specific context ofr InputEditModal
Elm stories for elm and react are generally configured to show the modal
in its opened state immediately. This is not a very realistic way to use
the modal so story had to change slightly.

When setting up state for the modal for InputEditModal we are setting
the default focusable id in the state to focus on the first field input.

This happens when the modal context is being set.