feat: Add InputEditModal variant for Elm Modal (#325)
* feat: Add InputEditModal variant for Elm Modal

Create a reusable accessible title function

This acts as a utility function that ensures an ariaLabelledBy id lives
on the parent div to the heading child

Add helper function for ariaDescribedBy.

This function adds the ariaDescribedBy to the id of the containing div
which surrounds the instructive text or subtext of the modal body.

* Expose function to set defaultFocusableId

A consumer may want a particular element in their view to be focused by
default.
This function lets the user set the id for the default focusable
element.

* Add button to reopen input edit modal

* Set up modal stories to set specific context ofr InputEditModal

Elm stories for elm and react are generally configured to show the modal
in its opened state immediately. This is not a very realistic way to use
the modal so story had to change slightly.

When setting up state for the modal for InputEditModal we are setting
the default focusable id in the state to focus on the first field input.

This happens when the modal context is being set.