fix: Focus lock fix when confirm button disabled (#358)
BREAKING CHANGE: Some exposed functions in Modal.Presets.Confrimation
have been renamed.

The LastFocusableElement may be disabled and therefore not focusable, in
this case we attempt to fall back to the FirstFocusableElement.

This is specifically for when the modal opens and during focus-lock.

Allow for disabled buttons to not receive ids

the confirm button when disabled will not be given an ID.

This is because Modal tries to focus it when it is a
LastFocusableElement. The Browser.focus result returns as Ok because it finds
the element but in reality it was not focused.

The confirmation modal by default sets the confirm button as the 'last
focusable element'. This is important for the focus lock implementation.

The confirm button can be set as disabled and in this case the button
will no longer receive focus and therefore not technically be the last
focusable element.

In this case we want to promote the footer dismiss button to be the last
focusable element whilst confirm is disabled and inverse once the
confirm is not disabled.