chore: Bump elm-storybook
This allows elm programs in storybook to leverage ports and execute
javascript.

Creating elm views that rely on ports brings up questions around how
this dependency is communicated to consumers.

Ongoing discussions are being had with the design systems team around
documentation.

The use of ports as it stands should ideally be used as a performance
enhancer to a problem not able to be solved with Elm alone.