feat: Use ports to dynamically adjust the select input on change
The idea here is to send a port to javascript land when the input
receives focus to observe the sizer node and adjust the input width
accordingly.

To rezise the input in pure elm requires constant dom queries per
keydown. On an empty page this might be very quick but on a large page
filled with dom nodes the lookup is usually very slow.

The result is that the width of the input always lags behind the user
input which makes for a not very nice user experience.

The ports will be used to create mutation observers and dynamically
add width to the select input during user input.

Write port glue for dynamic width

This code registers a mutation observer, calculates the width of the
input and disconnects the mutation observer.