Revert "feat: remove hover states for Tag"
Turns out these are still required even though they are not in Zen UI Kit.

I think we really need a prop for whether the Tag needs a hover state or is clickable, but
that's out of scope for this PR. This is a bit of a compromise. The non-hover state has
the border removed, but the border shows when hovered (so as to not break backwards
compatibility.

This reverts commit a8f58e4d5cc3da9cd2b61c747a506c91822a18f7.

I think we n