feat: Update border width and color + make borders invisible where appropriate (#408)
* fix: replace `border-radius: transparent` with token

the regex:

    find| border-color: transparent
    replace| border-color: border-color: $kz-border-borderless-border-color

no new imports were necessary

* fix: use transparent border token in Tag

* feat: remove hover states for Tag

The Zen spec does not have any hover states for Tag where
the border becomes visible

* feat: remove border visibility from all Tag variants

* feat: make Guidance block border transparent

* feat: give Guidance block round borders

* feat: uplift Toggle Switch to Zen borders, shadows and colours

The Toggle Switch is now fully Zenified *except* for the focus ring.
I removed the .freemium style class as it is now identical. However, I've left in
the prop the React/Elm component, as it still might be being used.

* Use tokens for border-width instead of local vars

This also updates VerticalProgressStep's styles, since they were depending on one of the local vars in Button.

* Use border width tokens for Dropdown

* Use border width tokens for MenuList

* Use border width tokens for remaining components

* Revert "feat: remove hover states for Tag"

Turns out these are still required even though they are not in Zen UI Kit.

I think we really need a prop for whether the Tag needs a hover state or is clickable, but
that's out of scope for this PR. This is a bit of a compromise. The non-hover state has
the border removed, but the border shows when hovered (so as to not break backwards
compatibility.

This reverts commit a8f58e4d5cc3da9cd2b61c747a506c91822a18f7.

I think we n

* soften the border color hover states

we now go from invisible to visible border when hovering, so
I have used my judgement and lighten the colour so the
change isn't as drastic. Let's review in Design Review.

* Fix focus state on Select

* Use correct borders and colors for checkboxes

* Update radio colours

* Make Guidance Block story use Stone background

* Fix Elm Select scrollbar bug

* Fix Elm TextField pop out bug on focus

* Fix ToggleSwitch focus bug

* Fix Tooltip arrow border color bug

* Add focus border color to TextAreaField

* Add old border width var back for safety

Co-authored-by: Michael Bylstra <mbylstra@gmail.com>
Co-authored-by: Michael Bylstra <michael.bylstra@cultureamp.com>