Add createAriaHider implementation
This is cribbed from reach-ui which implements the same pattern for its
Dialog component. The createAriaHider function essentially take a
reference node for the modal and then add `aria-hidden=true` attributes
to every direct descendant of the body when it’s called, and then
returns a function to undo that behaviour when its called.