Disable pointer-events within <label>
There are strange behavioural issues on iOS with enabling checkboxes
or radio buttons consistently by clicking on the visible "input" itself.

I say "input" because the input is really invisible and hidden behind
a div that represents its state, and something about that interaction
here with iOS touch/click behaviour means that checkboxes/radio buttons
sometimes do not respond. Disabling pointer events for everything inside
the <label> elements for each of those means that it all works as
expected.

I’ve done this at the component level rather than in the primitives to
avoid any complications from this behaviour leaking out beyond these
components. The downside to using this rule is that we can’t
differentiate pointer interactions (like :hover) _within_ the <label>
itself, but we’re not currently doing that and it’s unlikely we will.