fix: Checkbox and radio button inconsistent behaviour on touch devices (#550)
* Add browserFeature util as a common place for feature detection

* Only enable hover styles on non-touch devices

The `:hover` state on the label creates an intermediate "touchable"
state that activates the hover style _without_ effecting the touch down
to the underlying `<input>`.

This means that the Radio/CheckboxField components were more difficult
to select on iOS (and probably other touch devices) than expected.

* Remove tap highlight from labels

* Use @media(hover) for removing :hover issues

This feature is pretty well supported now, and in browsers that don’t
have support (this is only IE11) we will just not get hover styles that
are quite as nice for these components.

* Disable pointer-events within <label>

There are strange behavioural issues on iOS with enabling checkboxes
or radio buttons consistently by clicking on the visible "input" itself.

I say "input" because the input is really invisible and hidden behind
a div that represents its state, and something about that interaction
here with iOS touch/click behaviour means that checkboxes/radio buttons
sometimes do not respond. Disabling pointer events for everything inside
the <label> elements for each of those means that it all works as
expected.

I’ve done this at the component level rather than in the primitives to
avoid any complications from this behaviour leaking out beyond these
components. The downside to using this rule is that we can’t
differentiate pointer interactions (like :hover) _within_ the <label>
itself, but we’re not currently doing that and it’s unlikely we will.

* Scope pointer-events disable to inputs

* Remove JSX from changelog

Gatsby tries to evaluate this and fails to build.