feat: allow arbitrary Html in TextField description fields (Elm & React) (#671)
* feat: add descriptionHtml prop to Elm TextField

This allows arbitrary Html to be optionally used for the message. Currently
it is restricted to just strings. The use case was putting a link in
the description.

* feat: loosen restriction of TextField message content (React)

Previously it only allowed strings, now it allows arbitrary Html (React.ReactNode)

* feat: change descriptionHtml from single element to list

This is more like Elm's `div [] []`  convention, and is more convenient to use
as you don't need to think up a wrapping element.

* feat: add stories for Elm description and descriptionHtml TextField

* feat: add "Default with HTML description" story for React TextField

* feat: loosen type of description in TextAreaField

to match TextField