feat: Add focus support to the Button components (#662)
* Give the ability to imperatively call `focus` the two `Button` components.

![image](https://user-images.githubusercontent.com/4495057/87897327-53e67380-ca8e-11ea-9764-9ca1a815a898.png)

* Forward down any `aria-*` props down to the two `Button` components. Edit: support has been updated to allow for all kebab cased props.

### Motivation

I'm going through perf-ui at the moment, and making some a11y improvements. One of those improvements is for a dropdown menu, which uses this kaizen button. When the user presses escape, and we've closed the dropdown menu, we need to place the focus back on the original button.

The menu in perform, under "Goals":

![image](https://user-images.githubusercontent.com/4495057/87897513-f7378880-ca8e-11ea-872e-f12bb5a647d4.png)

I realised after the fact, that we already have a button dropdown in kaizen, and I should have just replaced the perform-ui code. But, I think the updates will still be handy at some point.