feat: point $ca-default-font-xxx to $ca-inter-font-xxx
BREAKING CHANGE

The breaking change is in only in the most legacy parts of Kaizen (`ca-` scss mixins and variables)
that are only used in the Big M repo.

For all other repos (non Big M), this won't affect them, so this release can be blindy
upgraded without testing. However, if you're unsure you can just do a search
for `ca-` in scss files in your repo, and if there are no results, no need to
worry about the major update

If you are bumping the Big M repo with this release, please PM michael.bylstra@cultureamp.com
for more details.

Description of changes:

- This removes all traces of Open Sans from Kaizen, and
makes Inter the default for $ca-default-font-xxx
- the $ca-default-font-xxx vars are legacy (in both files) and should be
removed eventually, but this is far easier (not a breaking change, codewise)
The nice thing is that the name $ca-default-font-xxx is still applicable, so
there's no need to change the name like we needed to do for $ca-font-ideal-sans!