feat: point $ca-default-font-xxx to $ca-inter-font-xxx
This removes all tracks of Open Sans from Kaizen, and
makes Inter the default for $ca-default-font-xxx

the $ca-default-font-xxx vars are legacy (in both files) and should be
removed eventually, but this is far easier (not a breaking change, codewise)
The nice thing is that the name $ca-default-font-xxx is still applicable, so
there's no need to change the name like we needed to do for $ca-font-ideal-sans!

The consuming repo (hint hint) will just need need to bump Kaizen, and hopefully
most of the Open Sans usages will dissapear.