feat: TitleBlockZen. Add a  variant for the badge (#826)
# Objective
Include a `default` variant for the badge that the component display for survey status.

# Motivation and Context 
In #team_apollo we are currently working on Evaluation cycles admin dashboards. In one of the pages that display feedback requests information we want to display the `due date` of the current Evaluation Cycle.
Since we only allowed bags for `live` and `draft` status, thought about adding a status that will use the `default` variant of the `Tag` component. 
In our case will be used, as mentioned above, for `due date` .

# Screenshots (if appropriate)
![image](https://user-images.githubusercontent.com/3228237/96280122-fd65d400-0fad-11eb-94f7-f77a7223fba6.png)

![image](https://user-images.githubusercontent.com/3228237/96280213-153d5800-0fae-11eb-82df-8cc194b095f4.png)

📹 : http://g.recordit.co/BKZ7DMGU9W.gif

# Checklist
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
[x] If this contains visual changes, has it been reviewed by a designer? 
[ ] I have considered likely risks of these changes and got someone else to QA as appropriate
[x] I have or will communicate these changes to the front end practice