feat: Allow TitleBlockZen to collapse when no navigation (#810)
* Remove unused var

* Add collapseNavigationAreaWhenPossible prop

This allows a consumer to (optionally) collapse the space usually used
by the navigation tabs in the case there are no navigation items.

* Handle collapsed nav on smaller screens

Main change here to the logic that previously used CSS to hide the `navigationTabScrollerContainer` on smaller screens. Instead of hiding with CSS, we allow it to be visible all the time, but only set its overflow properties when the `.navigationTabs` parent class is applied. This should be fine because the overflow properties only applied when that class was there before. A consequence of this is that we’d potentially have shown the shadow edges in situations we don’t want, so we now explicitly avoid rendering them when `collapsed` is true.

* Add clearer example stories for collapsed variants