fix: Navigation action buttons from being unnecessarily recreated (#807)
TitleBlockZen action buttons from being unnecessarily recreated. To do this, I've given proper keys to the primary and secondary buttons array.

The keys before were derived from the array index. This was resulting in a flaky cypress in perform-ui. We were trying to click the `primaryAction` button in the `TitleHeaderZen`. Every so often, the `click()` would fail, due to the element having disappeared from the dom. The cause being, in that short gap, the action item had been replaced with an exact replica.

I also noticed in the `ZenNavigationBar`, some menu items were getting recreated on every render, due to `key={uuid()}` on some of the components. I vaguely also remember having some issues trying to click on the top navigation bar in our cypress tests. I can't remember what the error was, but it sounds like this was likely the cause. Unfortunately I needed to use the array index, against react best practices, because the menu items have no unique identifier. To fix this, we'd need to create a breaking change, which will force all consumers to add a unique key for each menu item.