feat: Add tooltip to the Table header (#859)
# Objective
Adds a `tooltipInfo` property to the `Table` component. Using this prop will wrap the cell in a tooltip, and add a blue info icon to the left hand side.

The problem with simply wrapping the heading cell with a `Tooltip`, is that it ruined a lot of the flex styling. To overcome this, I needed to make some minor refactors to `TableHeaderRowCell`.

# Motivation and Context 

On the new Evaluation Cycle dashboards work, we needed to include a tooltip, to provide more context on the table header column name. We've had some confusion from customers in regards to when a request is "from" someone, vs "for" someone. We have very little horizontal realestate on this table, so a Tooltip was our option of last resort.

# Screenshots

![image](https://user-images.githubusercontent.com/4495057/98763638-93b2db80-242e-11eb-83f4-b1e5e0477032.png)

# Checklist
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
* [x] If this contains visual changes, has it been reviewed by a designer?  - Approved by Stuart
* [X] I have considered likely risks of these changes and got someone else to QA as appropriate
* [X] I have or will communicate these changes to the front end practice