feat: Set the linting max-warnings to 0 (#878)
# Objective
Set the linting max-warnings to 0. If the warning doesn't apply, the developer can simply add a `eslint-disable-next-line` comment.

# Motivation and Context 
Now that our warning count is back down to zero, this will ensure that the list of warnings doesn't climb up again. When a developer sees a large list of warnings, the messages start to become less meaningful.

# Checklist
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
[ ] If this contains visual changes, has it been reviewed by a designer? - NA
[X] I have considered likely risks of these changes and got someone else to QA as appropriate
[X] I have or will communicate these changes to the front end practice