feat: Allow for the modal confirmation button to be optional (#882)
# Objective
Allow for the modal confirmation button to be optional

# Motivation and Context 
I was requested to show a cautionary modal, to let admins know that they can't view their own reviews. The designs only have one button.

# Screenshots (if appropriate)

<img width="738" alt="image" src="https://user-images.githubusercontent.com/4495057/101001810-b494c980-35b3-11eb-9b13-c848d80643be.png">

# Side updates

Export props for the Popover and SplitButton components. Sometimes in perform, we add wrappers around the kaizen components to enhance the functionality. The common use case being linking them to react-router, or automatically assigning the translation strings. Exporting the props in kaizen, then allows for the wrapper component to take in the same props as its child.

# Checklist
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
[X] If this contains visual changes, has it been reviewed by a designer? 
[X] I have considered likely risks of these changes and got someone else to QA as appropriate
[X] I have or will communicate these changes to the front end practice