Use our existing SASS variables to toggle tokens
- Define our current tokens as CSS custom properties on `:root`
- Define our rebrand tokens as CSS custom properties on `.brand2021`
- Redefine the existing SASS variables to be `$kz-color-cluny-500: var(--kz-color-cluny-500);` etc
- Create new SASS variables for `$kz-color-cluny-500-default` which has the hex value for the token's default (not the rebrand)

Why?

This allows us to use a `.brand2021` class on the body to toggle all tokens for the page, allowing us to more easily QA and release with care.

What about IE11? IE11 users will be stuck on the default token set (so even with the feature flag won't be updated).

What else

- [ ] We'll need to update places where `add-tint` `add-shade` and `add-alpha` are used to now reference the `$kz-{token_name}-default` variables. These mixed colours won't be toggleable using the class.
- [ ] We'll need to fix a pinned dependency in title-block-zen