Add UI Kit sticker sheets to a bunch of stories
For every sticker sheet in UI Kit for which there is a corresponding
React component with stories, the sticker sheet is now added as a figma
design on those stories.

I haven't dived into make each individual story target an individual Figma
component, but that could easily be added by people as they work on stories.