docs: Embed UI Kit Figma designs in Storybook (#992)
* Add storybook-design-addons dependency

...and include it in our storybook builds

* Add a helper for setting figma parameters on a story

Is there a better place I should be putting this?

* Add figma embeds for all button stories

For any story with a matching element in UI kit, I've included the exact
link. The rest will fall back to showing the stickersheet as a whole.

* Add UI Kit sticker sheets to a bunch of stories

For every sticker sheet in UI Kit for which there is a corresponding
React component with stories, the sticker sheet is now added as a figma
design on those stories.

I haven't dived into make each individual story target an individual Figma
component, but that could easily be added by people as they work on stories.