rename @cultureamp/elm-storybook imports to just elm-storybook
This was confusing - it made it seem like the code was coming
from a node-package repo, but actually it's just coming from
the cultureamp/elm-storybook repo. There is actually an official
npm release (just `elm-storybook`, not `@cultureamp/elm-storybook`)
but it's not being used directly in this repo, as the latest
version was not released there.