fix: Make dividers bleed to edge of dropdown
BREAKING CHANGE: This includes a potential visual breaking change, since it restructures the way margins and padding work. If any consumers are using the Menu in non-standard ways, such as creating a section at the top of the dropdown that bleeds all the way to the edges by using a negative margin to override the dropdown container's padding, those could break.

Under the new model, instead of the container using padding to add space, the *content* is responsible for setting its own margins. (This is handled within the component, so consumers don't have to do anything - but as mentioned above, they could run into problems if they had been overriding either the margin or padding.)