Sibling CSS variable tokens and theme switching.
* Generate new token files containing CSS variable functions
* Generate a css/default-theme.css containing the variables of the default theme, along with css/heart-theme.css and css/zen-theme.css
* Add ThemeManager class to support subscribing to and switching the current theme
* Expose a React context wrapper around ThemeManager, with supporting hook: `useTheme`.

These code changes we're migrated from an older PR, #967.
It was migrated using `git checkout jasono/css-var-token-switching -- packages/design-tokens`, so the precise (but brief) commit history from that branch was lost.
The reason why cherry pick wasn't used is because there we're only a few commits related to design-tokens changes, and some had been mixed with changes to other directories too.

Cross posting from slack for more context:
We’re going to change the release path of the new tokens to make the migration of consumers and release of components much easier.
Rather than making a breaking change of each $kz-blah: #ff0011 variable, we’ll initially introduce new sibling variables that contain the CSS variable version of the same thing, like $kzvar-blah: var(--kz-blah) , which means that consumers will be able to update to the new design tokens version without any issues.
In the work done prior where we wanted this breaking change to happen, this was much more trouble than it was worth, mainly because we would have to make massive PRs which would be difficult to verify and review.
This way, consumers can update kaizen dependencies and design tokens asynchronously.