fix: Remove CSS properties when theme is Zen
## Context
We have the need to use different tokens depending on which theme is active. The most frictionless way we identified to do this was to allow for usages like: `var(--kz-color-stone-500, $kz-color-wisteria-800)`, which says: If there exists a CSS variable --kz-color-stone-500, use it, otherwise use wisteria 800. In effect, if we don't set any CSS variables when Zen is active, this means we have stone 500 on Heart theme, and wisteria 800 on Zen theme.

Great Success.